backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
PyMuPDF==1.24.10
groq==0.8.0
openpyxl==3.1.5
python-docx==1.1.2
python-multipart==0.0.12
